refactor(package): change package name to org.glpi.api
