+++ app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 26
     buildToolsVersion "26.0.2"
     defaultConfig {
-        applicationId "org.flyve.glpiproject"
+        applicationId "org.glpi.glpiproject"
         minSdkVersion 16
         targetSdkVersion 26
         versionCode 1
