+++ app/src/main/java/org/glpi/glpiproject/addItemExamplePayload.java
@@ -23,41 +23,27 @@
 *  --------------------------------------------------------------------
 */
 
-package org.flyve.glpi.utils;
+package org.glpi.glpiproject;
 
-import android.util.Base64;
+import java.util.List;
 
-import java.io.UnsupportedEncodingException;
+public class addItemExamplePayload {
 
-public class Helpers {
+    public List<data> input;
 
-    private Helpers() {
+    public addItemExamplePayload(List<data> input) {
+        this.input = input;
     }
 
-    public static String base64decode(String text) {
-        String rtext = "";
-        if(text == null) { return ""; }
-        try {
-            byte[] bdata = Base64.decode(text, Base64.DEFAULT);
-            rtext = new String(bdata, "UTF-8");
-        } catch (UnsupportedEncodingException e) {
-            e.printStackTrace();
+    public static class data {
+
+        public String name;
+        public String serial;
+
+        public data(String name, String serial) {
+            this.name = name;
+            this.serial = serial;
         }
-        return rtext.trim();
     }
 
-    public static String base64encode(String text) {
-        String rtext = "";
-        if(text == null) { return ""; }
-        try {
-            byte[] data = text.getBytes("UTF-8");
-            rtext = Base64.encodeToString(data, Base64.DEFAULT);
-            rtext = rtext.trim().replace("==", "");
-        } catch (UnsupportedEncodingException e) {
-
-            e.printStackTrace();
-        }
-
-        return rtext.trim();
-    }
 }
