+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.flyve.glpiproject.MainActivity">
+    tools:context="org.flyve.glpiproject.org.glpi.glpiproject.MainActivity">
 
     <TextView
         android:layout_width="wrap_content"
