+++ glpi/src/androidTest/java/org/glpi/ExampleInstrumentedTest.java
