+++ glpi/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.flyve.glpi" />
+    package="org.glpi.api" />
