+++ glpi/src/main/java/org/glpi/api/ServiceGenerator.java
@@ -23,25 +23,34 @@
 *  --------------------------------------------------------------------
 */
 
-package org.flyve.glpiproject;
+package org.glpi.api;
 
-import java.util.List;
+import java.util.concurrent.TimeUnit;
+import okhttp3.OkHttpClient;
+import retrofit2.Retrofit;
+import retrofit2.converter.gson.GsonConverterFactory;
 
-public class deleteItemExamplePayload {
+public class ServiceGenerator {
 
-    public List<data> input;
+    private static final int CONNECT_TIMEOUT = 180;
+    private static final int READ_TIMEOUT = 60;
+    private static final int WRITE_TIMEOUT = 60;
 
-    public deleteItemExamplePayload(List<data> input) {
-        this.input = input;
+    protected Retrofit retrofit;
+
+    protected void start(String url) {
+
+        OkHttpClient okHttpClient = new OkHttpClient().newBuilder()
+                .connectTimeout(CONNECT_TIMEOUT, TimeUnit.SECONDS)
+                .readTimeout(READ_TIMEOUT, TimeUnit.SECONDS)
+                .writeTimeout(WRITE_TIMEOUT, TimeUnit.SECONDS)
+                .build();
+
+        retrofit = new Retrofit.Builder()
+                .baseUrl(url)
+                .client(okHttpClient)
+                .addConverterFactory(GsonConverterFactory.create())
+                .build();
+
     }
-
-    public static class data {
-
-        public String id;
-
-        public data(String id) {
-            this.id = id;
-        }
-    }
-
 }
