+++ glpi/src/main/java/org/glpi/api/query/GetAllItemQuery.java
@@ -23,14 +23,16 @@
 *  --------------------------------------------------------------------
 */
 
-package org.flyve.glpi.query;
+package org.glpi.api.query;
 
 import android.content.Context;
-import org.flyve.glpi.R;
+
+import org.glpi.api.R;
+
 import java.util.HashMap;
 import java.util.Map;
 
-public class GetSubItemQuery {
+public class GetAllItemQuery {
 
     private Boolean expandDropdowns;
     private Boolean getHateoas;
@@ -38,13 +40,15 @@ public class GetSubItemQuery {
     private String range;
     private String sort;
     private String order;
+    private String searchText;
+    private Boolean isDeleted;
     private Context context;
 
     /**
-     * Get all the parameters available to work with the sub item endpoint
+     * Get all the parameters available to work with the all items endpoint
      * @param context
      */
-    public GetSubItemQuery(Context context) {
+    public GetAllItemQuery(Context context) {
         this.context = context;
     }
 
@@ -78,6 +82,14 @@ public class GetSubItemQuery {
 
         if(order!=null) {
             map.put("order", order);
+        }
+
+        if(searchText!=null) {
+            map.put("searchText", searchText);
+        }
+
+        if(isDeleted!=null) {
+            map.put("is_deleted", isDeleted.toString());
         }
 
         return map;
@@ -134,6 +146,22 @@ public class GetSubItemQuery {
      */
     public void setOrder(Order order) {
         this.order = order.name();
+    }
+
+    /**
+     * (default NULL): array of filters to pass on the query (with key = field and value the text to search)
+     * @param searchText String
+     */
+    public void setSearchText(String searchText) {
+        this.searchText = searchText;
+    }
+
+    /**
+     * (default: false): Return deleted element.
+     * @param deleted boolean
+     */
+    public void setDeleted(Boolean deleted) {
+        isDeleted = deleted;
     }
 
     /**
