+++ glpi/src/main/java/org/glpi/api/response/InitSession.java
@@ -23,34 +23,20 @@
 *  --------------------------------------------------------------------
 */
 
-package org.flyve.glpi;
+package org.glpi.api.response;
 
-import java.util.concurrent.TimeUnit;
-import okhttp3.OkHttpClient;
-import retrofit2.Retrofit;
-import retrofit2.converter.gson.GsonConverterFactory;
+import com.google.gson.annotations.SerializedName;
 
-public class ServiceGenerator {
+public class InitSession {
 
-    private static final int CONNECT_TIMEOUT = 180;
-    private static final int READ_TIMEOUT = 60;
-    private static final int WRITE_TIMEOUT = 60;
+    @SerializedName("session_token")
+    private String sessionToken;
 
-    protected Retrofit retrofit;
+    public String getSessionToken() {
+        return sessionToken;
+    }
 
-    protected void start(String url) {
-
-        OkHttpClient okHttpClient = new OkHttpClient().newBuilder()
-                .connectTimeout(CONNECT_TIMEOUT, TimeUnit.SECONDS)
-                .readTimeout(READ_TIMEOUT, TimeUnit.SECONDS)
-                .writeTimeout(WRITE_TIMEOUT, TimeUnit.SECONDS)
-                .build();
-
-        retrofit = new Retrofit.Builder()
-                .baseUrl(url)
-                .client(okHttpClient)
-                .addConverterFactory(GsonConverterFactory.create())
-                .build();
-
+    public void setSessionToken(String sessionToken) {
+        this.sessionToken = sessionToken;
     }
 }
